pyproject.toml
[tool.poetry]
name = "gabber"
version = "0.1.0"
description = ""
authors = ["R. Miles McCain <mccain@stanford.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
click = "^8.0.1"
python-dateutil = "^2.8.2"
loguru = "^0.5.3"
tqdm = "^4.62.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gabber = "gabber.client:run"